Modified the grouping of characters
Rather than grouping the characters for the key naturally have now gone
for the ease of observation. They are now grouped by first, seconf etc.
as first look, seacond look. This seems to match more the way an expert
approaches the problem and works with the capability of the key.